Update and rename playlist_stack.py to playlist_stack.c
This project uses a stack to create a playlist that operates on the 
last in, first out principle. Songs are played one at a time, and we 
also created a function that allows us to remove songs from the 
stack. The code can also indicate whether the stack is full or 
empty when adding or playing a song from the stack. Every 
song has a set amount of time allotted to it. When a playlist is 
created, it shows the total time required to hear every song.