Rename departmentalstore_array.py to departmentalstore_array.c
In this project, we created a department store where customers 
could purchase everything they wanted, including fruits, 
electronics, clothing, and medicines. We have designed several 
functions for certain choices, and once a user selects one, he is 
given the opportunity to purchase a variety of items. The final 
step displays a cart with everything the user chose.