Update and rename mailbook_linkedlist.py to mailbook_linkedlist.c
Mail-book is based on singly linked list. Mail book has four 
functions. The first option is "add an email," where the user 
enters the name and email address he wants to store. The second option is to delete an email. If the user types in the name of the recipient and the email is stored in the mailbook, the email is deleted. Additionally, we have the option of searching for emails by providing a person's name, and lastly, we have used file handling to show the mailbook in a .txt file