Update and rename tsp.py to tsp.c
The traveling salesman problem (also known as the TSP) asks, "Given a list of cities and the distances between each pair of cities, what is the shortest possible route that visits each city exactly once and returns to the origin city?"
Solution to the problem:(solved using graph data structure)
To solve the TSP using the Brute-Force method, you must first calculate the total number of routes and then draw and list all of them. Calculate the distance between each route and then select the shortest one—this is the best option.